# api/openapi.yaml
openapi: 3.1.0
info:
  title: NXT TIME PULSE Admin API
  version: 1.0.0
  description: Admin-gated API endpoints for GEO TIMESHEET - TIMEPULSE
servers:
  - url: /api
    description: API base path
security:
  - bearerAuth: []
paths:
...
  /assignments:
    get:
      summary: List project assignments
      tags: [Assignments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: query
          schema: { type: string }
        - name: projectId
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string, enum: [active, paused, ended] }
      responses:
        "200":
          description: Assignment list
    post:
      summary: Create assignment (critical/backdated changes queue approval)
      tags: [Assignments]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, projectId, roleOnProject, startDate]
              properties:
                userId: { type: string }
                projectId: { type: string }
                roleOnProject: { type: string }
                allocationPct: { type: number }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date, nullable: true }
      responses:
        "200":
          description: Assignment created
        "202":
          description: Pending approval
  /assignments/{id}:
    patch:
      summary: Update assignment (critical/backdated changes queue approval)
      tags: [Assignments]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                roleOnProject: { type: string }
                allocationPct: { type: number }
                startDate: { type: string, format: date }
                endDate: { type: string, format: date, nullable: true }
                status: { type: string, enum: [active, paused, ended] }
      responses:
        "200":
          description: Assignment updated
        "202":
          description: Pending approval
...
  /rates/{id}:
    get:
      summary: Get a rate card
      tags: [Rates]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Rate card
    patch:
      summary: Submit rate-card update request (maker-checker)
      tags: [Rates]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                payRate: { type: number }
                billRate: { type: number, nullable: true }
                currency: { type: string, minLength: 3, maxLength: 3 }
                effectiveFrom: { type: string, format: date }
                effectiveTo: { type: string, format: date, nullable: true }
                status: { type: string, enum: [pending, active, superseded, cancelled] }
                changeReason: { type: string, nullable: true }
      responses:
        "202":
          description: Rate change queued for approval
  /rates/resolve:
    get:
      summary: Resolve effective pay/bill rate by date and scope
      tags: [Rates]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: query
          required: true
          schema: { type: string }
        - name: projectId
          in: query
          required: true
          schema: { type: string }
        - name: date
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Resolved rate
...
  /financials/summary:
    get:
      summary: Financial summary (admin/manager only)
      tags: [Financials]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: from
          in: query
          required: true
          schema: { type: string, format: date }
        - name: to
          in: query
          required: true
          schema: { type: string, format: date }
      responses:
        "200":
          description: Summary data
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "Unauthorized" }
    Forbidden:
      description: Insufficient permissions (requires org:admin or org:manager)
      content:
        application/json:
          schema:
            type: object
            properties:
              error: { type: string, example: "Forbidden" }